Add java/lang/StringUTF16.newBytesFor to alwaysWorthInlining list
As part of #8830 we started using this API in several String
constructors which are quite performance sensitive. Rather than relying
on frequency information always being correct we simply force the
inlining of this small method which will exposed the allocation of the
array which we will store into the String object.

Closes: #8831

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>