Runtime compressed refs work
Undo an incorrect change in the monitor table (restore the original
code).

The existing code works in the normal builds because j9objectmonitor_t
and the lockword size are identical.

Mixed mode on X appears to work because the table is zeroed to start
with, and since X is little endian, if the stored value is smaller than
4Gb then there's no problem. If any J9ObjectMonitor instances were
allocated above the 4Gb bar, I expect things would fail.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>