Use OMR_GC_POINTER_MODE
Instead of specifying either OMR_GC_COMPRESSED_POINTERS or
OMR_GC_FULL_POINTERS, we will be using an enumeration to specify mode of
references in OMR.

OMR_GC_POINTER_MODE can be set to full, compressed, or mixed.

This is the first of several changes to migrate OpenJ9 to the new
option.

Signed-off-by: Andrew Young <youngar17@gmail.com>