Fix definingClass for StaticFieldVarHandle
For a StaticFieldVarHandle, definingClass should point to the owner of
the static field.
 
Example:
    public class P { public static String publicObject = "s"; }
    public class C extends P { }
    VarHandle vh =
    	MethodHandles.lookup()
    		.in(C.class)
    		.findStaticVarHandle(C.class, "publicObject", String.class);

In the above example, vh is an instance of StaticFieldVarHandle, which
extends FieldVarHandle. vh.definingClass should contain class P since it
defines the publicObject.

Currently, vh.definingClass is set to class C, which is not the owner of
the static field. This fix correctly sets the defining class for a
StaticFieldVarHandle.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>