Partial MethodType created by readObject can be used
If the serialized MethodType contains objects which have their own readObject
method, whether from extra items in the serialized form or from the return
type or argumens fields, then the bogus MethodType object created will be
accessible from the other readObject methods, with possibly incomplete type
information.

This commit adds copies of the arguments and return type fields for use in
serialization, and sets the actual type to "()V" until readResolve is called
and creates a useable MethodType.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>