Change stackWalkerMaySkipFrames to avoid AOT failures
Current implementation of `TR_J9SharedCacheServerVM::stackWalkerMaySkipFrames`
tries to avoid messages to the client by using a couple of calls to
`TR_J9ServerVM::isInstanceOf()`. However, under the covers, the call to
`TR_J9ServerVM::isInstanceOf()` will result into calls to
`TR_J9SharedCacheServerVM::getSuperClass()` which will check to see if
a SVM validation record for the given class is already present, which is
not. As a consequence, the SVM validation check will fail the AOT
compilation.
The solution implemented in this commit is more or less a workaround:
when SVM is used during an AOT compilation, do not try to optimize
`TR_J9SharedCacheServerVM::stackWalkerMaySkipFrames`, but rather send a
message to the client to find the answer.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>