Remove final from fields being set in native methods
A final field should not be modified after initialization. The two final
fields mentioned below are modified after initialization in native code
so they no longer satisfy the conditions of the final keyword. Hence,
the final keyword is removed for the two fields.

MethodHandle.kind is set in the following native method:
1. Java_java_lang_invoke_PrimitiveHandle_lookupMethod

VarHandle.modifiers is set in the following native methods:
1. Java_java_lang_invoke_FieldVarHandle_lookupField
2. Java_java_lang_invoke_FieldVarHandle_unreflectField

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>