Reverse Throwable writableStackTrace boolean field logic
The field to control Throwable writeable stack trace
enableWritableStackTrace has a default value true. The field is
transient (not included in the serialized form) as per the
specification. This logic doesn't work with Google gson, as gson doesn't
restore the field value after serialization. Rename the field to
disableWritableStackTrace, which is false by default, and reverse the
logic.

Also remove checks if the disableWritableStackTrace field exists, which
were for Java 6.

Fixes #8760

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>