Static/special resolution paths should only perform interpreted dispatch
The AMD64 JIT's static and special unresolved paths perform the
resolution and then conclude by either dispatching to the interpreted
method or the compiled body.

However, these resolution helpers overwrite many of the linkage registers
used in JIT private linkage such that performing a dispatch to a compiled
method is incorrect.  There is a comment to that effect before each
function that seems to be unheeded.  The stack is set up to allow
interpreted dispatch only.

The most straightforward way to fix this is to always dispatch to
the interpreted method from these helpers.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>