Cut number of SharedCache_getClassChainOffsetInSharedCache messages
Many such messages are issued from:
TR_J9JITServerSharedCache::getClassChainOffsetOfIdentifyingLoaderForClazzInSharedCache()
This commit reduces the occurences of such messages by caching the mapping
{j9class -> classChainOffsetOfIdentifyingLoaderForClazz}
JITServer already maintains a mapping {j9class -> ClassInfo} so the work
here involves extending the struct `ClassInfo` with an additional field.
One problem that needs to be taken care of is that
`classChainOffsetOfIdentifyingLoaderForClazz` may not be available when
`ClassInfo` for a j9class is cached. Thus, we need an additional check
against the cached value. If that turns out to be 0 we need to issue a
message to the client to find the proper value.

Closes: #8959

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>