Optimize IProfiler fan-in info serialization
Previously, serializing `TR_IPMethodHashTableEntry` to
`TR_ContiguousIPMethodHashTableEntry` required memsetting
the contiguous entry to 0 on the client, so when
the next element in the serialized array of callers is NULL,
server would know that it reached the end.

This commit adds `_callerCount` field to the contiguous entry,
so that client doesn't have to perform a memset since server
is informed how many callers are present.

Also modify serialization function for method entries to remove
one round of copying.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>