Add query method for class whose instances are zero initializable
Define an isZeroInitializable method that indicates whether a class's
instances have fields that can be trivially initialized by zeroing
memory.

For prototype support of value types, a class with a value type field
that has not been inlined into the class must have that value type
field of its instances initialized with the default value for the value
type.  The implementation of isZeroInitializable is introduced to detect
that case.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>