Add valueType checks to monitor enter
The change updates the non-blocking monitor enter code to detect
valuetype receivers and return an appropriate error code. This means
that the `1 == rc` check that is currently being done will not work as
there is now more than one error code in the non-blocking enter case.

Instead the following check must be done:
`(monstatus < J9_OBJECT_MONITOR_BLOCKING)`

More error codes may be added in the future.

It is not necessarry to modify the monitorExit code in the case of
valuetypes because the monitorExit code will throw IMSE if the receiver
is not locked. So it is only necesarry to disallow valueType receiver
from being locked in the first place.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>

# Conflicts:
#	runtime/oti/ObjectMonitor.hpp
#	runtime/vm/ObjectMonitor.cpp