Fix VT jitnewValue slow path
Currently, the function returned from the slow path takes in 3 params,
one for j9vmthread and 2 booleans for class init and nonZeroTLH.
However, the thunk sequnce in [x|z|p|etc]nathelp.m4 only passes in the
j9vmthread. This means that when `slow_jitNewValueImpl` is called there
values for the two booleans are corrupt.

This change introduces a wrapper function that passes in the correct
values for the boolean args.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>