Change monitor enter return type
When the previous code returns an IDATA for monitor status. A check is
performed to determine if the monitor status is less than the blocking
enter return code (currently 4). This is broken because a valid object
can be a negative value (highly likely on 32bit platforms).

This PR changes the return type from a IDATA to a UDATA. Also, it fixes
the bytecode action order.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>