String table read barrier
Replacing Collector (Scavenge) specific code in a middle of StringTable
code, with more generic barrier API.

Besides, the original code was not fully correct, since it was using
ForwardedHeader variant that is used for Partial GC in Balanced, that
is not aware of:
1) self forwarded pointer.
2) large object parallel copy protocol

All other code is just renaming various ReadBarrier APIs, giving them
more generic name that serves the same purpose for any kind of Weak
(Clearable) Root: just updates the slot in case object is already
forwarded, but does trigger a copy (keep it alive), even if in Evacuate
space.

This code has several more suspicious spots needing further scrutiny:
1) referring directly to Metronome and
2) reading object slots directly without using Read Barrier API
(currently, probably benign because those are pre=tenured objects not
subject to movement, but obviously not future-proof),
but the change does the bare minimum to fix a known problem.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>