Flush caches for threads without VM access only
Now that cache flushing is simplifed
(https://github.com/eclipse/omr/pull/5041), and directly
accesses/modifies target's thread caches/remainders in context of
calling thread (that walks all threads), we should not call it for
threads that are actively running (have VM access).

This is a minimum we must do, but we also need to prevent threads that
are about to acquire VM access and race with the calling thread.

This is why we hold target's thread publicFlagsMutex, although we might
need extra mechanism to force target thread to take slow
acquire-VM-access pass (that will also acquire publicFlagsMutex and
serialize with caller thread). This is still subject to testing (all
this code is still inactive).

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>