Forced finalize fix for CS
When VM is shutting down, we force finalization of all
objects even those that have not died. We do so by moving the content of
unfinalized list to finalizable list.

If this occurs in a middle of CS cycle, we would incorrectly leave
unfinalized _priorList head pointer non-null, which would then be later,
 in clearable pass of that CS cycle, be treated as if objects are still
in unfinalized list, while they actually belong to finalizable list.

Another problem is calling startUnfinalizedProcessing which would have
been already done at the start of this cycle, hence overwriting existing
_priorList pointer and losing the real content of unfinalized list.

Fix is to iterate the head list directly and reset the pointer after the
pass.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>