Abort ILGen for non supported features in AOT
Constant Dynamic, Method Handle Constant, and Method Type Constant
 is not currently supported in AOT. Therefore, this
commit throws the appropriate exception in ILGen if this
condition is met. However, because all of these exceptions are a
TR::RecoverableILGenException, if the compiler is generating IL for an
inlined method, it will not abort the compile, but simply refuse to
inline that particular method.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>