Prevent increasing the scratchSpaceLimit on low virtual memory
When low virtual memory is detected, the JIT will decrease the
`scratchSpaceLimit` (i.e. maximum amount of memory a compilation
thread is allowed to use) to 64 MB. This is done only if the current
scratchSpaceLimit is still at the default value of 256 MB.
However, a recent commit has changed the default scratchSpaceLimit
under `-Xtune:virtualized` option to 30 MB. Thus, we may actually
increase the scratchSpaceLimit from 30 MB to 64 MB when low virtual
memory is detected.
This commit adds a check that prevents the increase of the
scratchSpaceLimit under low virtual memory conditions.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>