Use _binaryBufferStart as reference for offset calculations
JITServer has been using getCodeStart() as the reference for offset
calculations when dealing with the patching involved in runtime assumptions.
However, this reference is dependent not only on the startPC but also
on the size of preprologue. This will create problems if we use the reference
before we generate the preprologue or have determined the preprologue
size. We use _binaryBufferStart to remove the preprologue size depedency.

Issue: #8786

Signed-off-by: Harry Yu <harryyu1994@gmail.com>