Fix known object table update at the server
When getIndexAt() is called, the passed-in parameter
objectReferenceLocation is not necessarily the reference
that’s stored in the known object table. It should not be
used to pass into updateKnownObjectTableAtServer()
to update the table at the server.

Pass the reference object location from the client to
updateKnownObjectTableAtServer() in getOrCreateIndexAt()
instead of the one passed in from the caller of
getOrCreateIndexAt().

Fixes #8952

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>