Post-store macro relying on barrier type
Post-store macros was incorrectly relying on GC allocation type, while
it should be relying on barrier type.

This was fixed for pre-store macros, not too long ago, but we missed to
fix for post-store.

This should somewhat reduce overhead of barriers in Optthruput GC policy
that in some cases (like array copy) were triggered for no good reason.

Also 'always' type was not properly handled to trigger pre-store.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>