Reduce number of VM_getClassFromSignature messages
JITServer employs a cache to reduce the number of
VM_getClassFromSignature messages, but this not entirely
efficient for AOT compilations because the AOT frontend
also performs some validation of the class found. If the
validation fails getClassFromSignature() return NULL
and the caching is not going to be performed at the server.
In this commit the client will call the TR_J9VM version of
getClassFromSignature() irrespective of the type of frontend.
The class that is found is sent to server which caches it
and then proceeds to perform the validation.

Issue: #9116

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>