Minimize includes in JitDump.cpp/hpp and reinstate extern "C"
We are reinstating the `extern "C"` qualifier here on the two functions
we expose because although not marking them as such still works it
depends on the include order, i.e. whether we include j9protos.h before
we declare the non-extern functions. This is fragile, so to avoid these
linker issues we reinstate the `extern "C"` qualifier here.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>