Disable upgrades/downgrades on jitdump compilations
When a jitdump recompilation is requested we want to reproduce the
exact same compilation which happened previously. This means we do not
want the normal JIT logic for upgrading or downgrading requested
optimization levels based on the state of the environment at the time
of the compilation request to affect our optimization level request.

We disable the path which upgrades/downgrades compilations if we
encounter a jitdump compilation request.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>