Drop the const qualifier on `J9Method*` in IlGeneratorMethodDetails
`J9Method* const method` is a constant pointer to a mutable `J9Method`.
This is not really a useful concept since we end up storing the
non-constant pointer in the class anyway! What I think was intended
here was to ensure that the data being pointed to be marked as
constant. That is represented by `const J9Method* method`. However,
making this change would require a cascading effect on many APIs
which is not feasible at this point. So instead we just drop the const
qualifier on the pointer.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>