Use a different cacheDir for the nameOption test
Use a non-default cacheDir for the
cmdLineTester_SCCommandLineOptionTests nameOption test. The test creates
and destroys the default cache name. When the shared cache is enabled by
default, on platforms where the default cache type is nonpersistent
(z/OS), the default cache in the default cacheDir may be in use by other
processes and cannot be destroyed.

Also remove the unnecessary javac commands, as the build.xml
pre-compiles the code.

[ci skip]

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>