Guard unloading class events and compilations with RWMutex
Use RWMutex to guard the access to the client session data
during the compilation to prevent the race condition
in accessing the caches related to the unloaded classes
at the server.`TR::CompilationInterrupt` exception is
thrown at the server if one compilation thread tries to do
a read access to the client session cache while another
compilation thread just attempted a write access to the same
client session data. It results in a `compilationFailure`
message along with statusCode=compilationInterrupted being
sent to the client.

Implements #9212

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>