Minimize contention between "finders" using SCC JCL helper
There are 2 issues with findSharedClass() in SCC Helpers being
synchronized:

A. For classLoader with isParallelCapable = true, the parallel 
class loading capability is lost.
B. In the case that there are multiple custom Classloaders setting 
the app ClassLoader as the parent, they all delegate to the app
ClassLoader first when finding classes. So all of them need to go 
through the synchronized method appClassloader.findSharedClass() 
one by one, which becomes a bottleneck.

The changes here are:
1. Change findSharedClass() as well as storeSharedClass() to be
non-synchronized.

2. Use a readWriteLock in SharedClassURLClasspathHelperImpl 
instead.

3. Do not enter classTableMutex and class segmentMutex in the 
native implementation of findSharedClass() in shared.c, as
entering these mutexes is essentially making findSharedClass()
synchronized in the native level.

4. Enter class segmentMutex only when necessary in 
hookFindSharedClass() and cacheMap::findROMClass().

5. Do not need to wrap callFindLocallyDefinedClass() inside class
segment mutex.

6. Change SharedClassURLHelperImpl.jarFileNameCache from a regular 
Set to a concurrent Set.

7. _identifiedClasspaths should be used inside _identifiedMutex.

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>