Send CHTable updates with compilation requests
In this commit the list of unloaded classes and the serialized version
of the CHTable updates are sent to the JITServer together with the
compilation request, rather than in a separate message. This has
two advantages:
(1) Fewer messages
(2) Serialization of compilation requests is sped up because it does
not have to wait for a round trip to ask for CHTable updates.
If caches are cleared (this includes the server version of CHTable)
then JITServer will still have to ask for the entire set of ranges
of unloaded classes and the entire CHTable (not just modifications
since the last message).
All these updates are processed in order based on the seqNo sent by
the client.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>