Handle OSR from inlined synchronized method
When a synchronized method exits through unhandled exception, it's
required by the specification to release the monitor before exits. The
VM unwind logic will handle this for synchronized methods that are not
inlined. For inlined ones, JIT inserts a synthetic handler to catch the
exception and do monitorExit and then rethrow the exception.

In case of OSR, the VM needs to be aware of those synthetic handlers to
resume execution correctly in the interpreter. Special catch type value
in the exception table entry is used for this purpose. Also, a new
helper for throwing exceptions from synthetic handlers is added.

This is the JIT side of the change.

see issue #9267

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>