Cache result of class chain validation
This commit introduces the ability to cache the result of the validation
of a class' class chain. The class chain validation validates the shape
of a loaded class; as such it does so using romclass comparisons.
Therefore, the shape can't change for the duration of the run.
Therefore, once a class has been validated (or failed to validate), that
result is true for the duration of the run.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>