Register assumptions for static final field modifications
This change serializes the list of final field folded classes
collected on the server during compilation and sends them over
to the client. Upon receiving the list of final field folded classes
from the server, the client checks for static field modifications
in the list of classes and fail the compilation if any of the classes
has its static field modified. If none has its static field modified,
then the client will proceed to create runtime assumptions and commit
them to the CHTable (done in JITClientCHTableCommit)

Issue: #9364

Signed-off-by: Harry Yu <harryyu1994@gmail.com>