Cache illegal final field modification at the server
Sends a list of classes that have J9ClassHasIllegalFinalFieldModifications
bit set to the server in the compilation request. The class cache
at the server is updated before the compilation. During the compilation,
the server only sends the message to query the value if the class is
not cached. If static field folding takes place in the
compilation at the server, the compilation is aborted at the client
during the CHTable commit (implemented in #9396)
if J9ClassHasIllegalFinalFieldModifications is set at the client
during the remote compilation.

Implements #9358

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>