Fix TR_J9ServerVM::canAllocateInlineClass
A recent commit for value type inlinine allocations changed the
implementation of TR_J9ServerVM::canAllocateInlineClass, but introduced
a bug since the classFlags of the ramClass are obtained only
if the class is not initialized.
This commit obtaines the value of the classFlags from the JITServer cache
in all cases, thus eliminating the bug and avoiding extra messages.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>