Incomplete tactical implementation for -XX:+PrintFlagsFinal
An incomplete implementation of VM options that produces the
-XX:+PrintFlagsFinal output for two specific options required by
the elasticsearch project. This code is expected to be temporary
until a more comprehensive solution can be architected. For this
reason, this change is very straight-forward and localized so
it can be easily removed.

The two options whose values will be printed are for
`-XX:MaxHeapSize=` and `-XX:MaxDirectMemorySize=` .

With this change, elasticsearch no longer NPEs on startup and,
in my local testing, all their tests pass on OpenJ9.

I tried to match, as closely as possible, the output Hotspot
generates when one uses -XX:+PrintFlagsFinal, which includes the
names used for types (e.g. size_t or uint64_t), the option names,
the option category (e.g. {product} for these two options), the
way the option is set (e.g. {ergonomic} or {command line} for
these two options), and the field widths produced by a JDK11
Hotspot build.

I added 3 command-line tests :

1) that two options (MaxHeapSize, MaxDirectMemorySize) are present
when -XX:+PrintFlagsFinal option is specified.

2) that -XX:+PrintFlagsFinal overrides earlier -XX:-PrintFlagsFinal.

3) that -XX:-PrintFlagsFinal overrides earlier -XX:+PrintFlagsFinal.

Fixes: #7764

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>