Leverage CHTable at the server to check if a class is extended
If a class has subclasses in the CHTable at the server,
return true in classHasBeenExtended(). If not or the class
info doesn't exist in the CHTable, check the ROMClass cache
next. `VM_classHasBeenExtended` is only sent when the class info
doesn't exist in the CHTable and the flag is not set in the
cache. `ResolvedMethod_getRemoteROMClassAndMethods` is sent when
the class data cannot be found in the CHTable or the ROMClass
cache at the server.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>