Deprecate TR_TraceCGMixedModeDisassembly with TR_TraceAll
This option is never really used in the field and the tracing is not
too useful either because the JIT compiler only ensures the bytecode
indices are valid at GC points, so mapping bytecodes to instructions is
rarely useful. In either way this tracing option is not needed as the
same infromation can be extracted via `TR_TraceAll` or `TR_TraceCG`.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>