Consolidate all trace register assignment options into TR_TraceRA
When investigating register assignment problems I have yet to encounter
a use case where we would not want to enable all RA tracing. In fact I
don't think many people are aware of the various sub-options you can
pass through to the `traceRA={}` regex. We simplify this logic here and
introduce a single option `TR_TraceRA` which encompasses all RA tracing.

We further take this opportunity to deprecate all APIs which are now
dead code as a result.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>