Disable GCR for hot compilations
GCR (guarded counting recompilations is a mechanism that is supposed
to upgrade cold or AOT compilations to warm opt level. In theory
it should work with any opt level, but due to a bug it seems that
GCR is incompatible with hot compilations (see #4548 for details).
To allow stress testing with hot compilations this commit disables
GCR for hot compilations. This should have no bearing on behavior
in production because the natural way of upgrading hot compilations
to scorching is through sampling.

Fixes: #4445 #8064

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>