Rewrite opt for combining consecutive byte loads
Changes were made inside the sequentialStoreSimplification. Despite the name
of the optimization, the changes are targeted at sequential loads.

Rewrote the optimization that identifies byte loads from conseuctive memory
addresses and replaces them with a wider load such as an iload or lload. The
new version greatly expands the number of tree patterns indentified by the
optimization. The new wide load optimization works with 2-8 consecutive byte
loads and in both big and little endian order.

The original version of the opt is still available under the env var
TR_UseOldSeqLoadOpt.

Moved sequentialStoreSimplificationGroup later in the optimization order so
that the trees are more likely to be in the right shape for the optimization.

Added tests to exercise the combining of sequential byte loads in the
sequentialStoreSimplification optimization.

Closes: #11250
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>