Don't remember historical thread names by default
There are use cases where a TraceContext object has a very long
lifetime, in which case it may not be reasonable to capture the
set of all names of each VM thread. This introduces API to control
whether historical thread names are captured (with the default
behavior being not to collect them). The traceformat utility uses
this new API to retain its current behavior.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>