Fix instanceOf's dynamicCache sequence on IBM Z
On IBM Z, for instanceOf node, we have a codegen optimization which
caches the encountered classes for which it failed all the inline tests
and has to call JIT helper to get the result. This cache is first
tested to avoid a call to JIT helper. There was a bug in case we have to
cache class of child when cast class child of the node is runtime
variable. We were loading and storing associated fields using two load
and store instructions. This causes a potential data race condition
which was exploited by highly parallel application where more than 500
threads were running same method and all were trying to update the
cache, leaving with the classes in the cache which were not related and
in the code it can later cause an incorrect result for instanceof.

Fixes: #9457

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>