VTs now use Qsignatures in class refs
Spec change:
VT must use Qsignatures in methodrefs, fields and now classrefs. This
means that when we are resolving classrefs we need to handle cases where
there is no signature (just name) and where there is a Qsignature. This
change has been made in the resolve classref code.

The rest of the changes are test related. Since there is no support for
VT with methodhandles, the method handles still look for 'L' signatures.
So as a work around the test generates a wrapper (generic) function that
accepts 'L' signatures and calls the real function with the 'Q'
signatures. The test cases call the generic version which then redirects
to the specific version.

For example:

```
  public static java.lang.Object makeValueGeneric(java.lang.Object,
java.lang.Object) throws java.lang.Exception;
    descriptor: (Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;
    flags: (0x0009) ACC_PUBLIC, ACC_STATIC
    Code:
      stack=2, locals=2, args_size=2
         0: aload_0
         1: checkcast     #58                 // class "QPoint2D;"
         4: aload_1
         5: checkcast     #58                 // class "QPoint2D;"
         8: invokestatic  #60                 // Method
makeValue:(QPoint2D;QPoint2D;)QFlattenedLine2D;
        11: areturn
```

This need for the generic wrapper has always been there but it is more
important now since VTs are only ever using Q signatures.

The increased usage of generic wrappers has exposed some bugs. Below is
a summary:
- Since MHs are not being used to directly call the make function we
cannot rely on the asType conversions the MHs would normally do, so the
types have to be exact in order to pass the checkcast test. Ie. use
longs instead of int, etc.
- The makeGeneric generator code was missing breaks in the case
statements
- VT return signatures were using 'L' instead of 'Q' signatures

A follow on PR will be created to address the new behaviour for
checkcast.



Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>