Resolve classref in checkcast if Qtype
Currently, if the receiver for a checkcast is NULL the VM does not
bother resolve the classref and the checkcast test passes. With VTs this
behaviour is not allowed as NULL cannot be assigned to VTs. VTs will
always have Qsignatures, therefore, if the classref oin a checkcast has
a qsignature it must be resolved, regardless of what the receiver is.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>