Fix incorrect types in value JProfiling quickTest block
Previously, the code for creating JProfiling quickTest blocks had an
error where the result of an lcmpeq node was being used as if it were a
long. This is invalid: compare nodes always result in an int, regardless
of the types of values being compared. This could result in erratic
behaviour, since both the optimizer and codegen generally assume that
the IL is typed correctly.

Signed-off-by: Ben Thomas <ben@benthomas.ca>