Reduce the scratch segment cache in JITClient from 16MB to 2MB
Remote compilations typically take longer due to network latency. This
increases the size of the compilation queue and, as a result, more
compilation threads can be activated. In order to keep the footprint
envelope low in the presence of many compilation threads we need to use
a smaller scratch segment cache. All remote compilation require less than
2MB at the JITClient so this seems like a good value.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>