Fix for 16bit load on Z with balanced GC policy (Arraylets)
When loading a 16bit value from memory while using the balanced
gc policy the Z code generator might generate a "LGH" rather then
a "LLH" instruction in some cases. This would result in an unintended
sign extend of the loaded value. This change will make the 16bit
loads work in a similar way as the 8bit and 32bit load operations
do when using arraylets.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>