Weak root Read Barrier for VLHGC
While strictly speaking VLHGC does not require require Read Barrier
(Copy-Forward is fully STW operation), stringComparatorFn can be used
both by Mutators and GC threads (in a middle of STW operation).

During comparison,
- when fetching a string from StringTable that has already been
forwarded, a forwarded variant needs to be fetched. If
- when fetching a string that has NOT been forwarded (since ST in not a
hard root), we do not copy-forward but use original variant

So, this definitely acts as a RB, and even if invoked from GC threads
during STW operation. Since the comparator method can be invoked from
Mutator as well, common denominator is to use RB API.

This was already addressed by
https://github.com/eclipse/openj9/pull/9330,
but it missed to preserve the functionality for VLHGC.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>