Add Peephole extension for handling literal pool register in catch block
We also take this opportunity to document the peephole. Although this
is not really a peephole optimization (as it is mandatory) it used to
belong in the OMR peephole phase, so we migrate it here instead. Rather
than finding a good place for this to live we currently leave it as is
since z9 support may be going away soon anyway.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>