Fix ilgen not clearing node flag when auto slot 0 is overwritten
TR_J9ByteCodeIterator::isThisChanged is supposed to check for a store to auto
slot 0. This allows TR_J9ByteCodeIlGenerator::loadAuto to clear the
nodeIsNonNull flag if the value in auto slot 0 has changed.

However, isThisChanged only handles astore bytecodes and not other types of
stores such as istore and lstore. This is causing issue #9381 where there is an
istore to auto slot 0 but the nodeIsNonNull flag is not cleared.

- Add cases in TR_J9ByteCodeIterator::isThisChanged to handle other types of
  stores such as istore and lstore.

Fixes: #9381

Signed-off-by: Ryan Shukla <ryans@ibm.com>